Deleted the unused files and merged files that are the same like the textfields files.
