+++ lib/SignUpLogin&LandingPage/sign_up_page.dart
@@ -5,12 +5,12 @@ import 'package:supabase_project/AuthService/snack_bar.dart';
 import 'package:supabase_project/CommonWidgets/dialogs/error_dialog.dart';
 import 'package:supabase_project/CommonWidgets/dialogs/loading_animation.dart';
 import 'package:supabase_project/CommonWidgets/loading_page.dart';
+import 'package:supabase_project/CommonWidgets/textfield.dart';
 import 'package:supabase_project/ConstantTexts/Theme.dart';
 import 'package:supabase_project/ConstantTexts/colors.dart';
 import 'package:supabase_project/SignUpLogin&LandingPage/login_page.dart';
 import 'package:supabase_project/SignUpLogin&LandingPage/setup_profile.dart';
-import 'package:supabase_project/SignUpLogin&LandingPage/textfield.dart';
-import 'package:supabase_project/SignUpLogin&LandingPage/user.dart';
+import 'package:supabase_project/ConstantTexts/user.dart';
 import 'package:supabase_project/buttons/sign_up_button.dart';
 import 'package:http/http.dart' as http;
 import 'dart:convert';
