--- lib/zNotUsedFiles/OldAPI.dart
@@ -1,38 +1,0 @@
-/*
-  Future<void> fetchAppliances1() async {
-    setState(() {
-      isLoading = true;
-    });
-
-    final prefs = await SharedPreferences.getInstance();
-    final userId = prefs.getString('userId');
-
-    if (userId == null) {
-      print("User ID is null. Cannot fetch appliances.");
-      setState(() {
-        isLoading = false;
-      });
-      return;
-    }
-
-    final url = Uri.parse(
-        "http://10.0.2.2:8080/getAllUsersAppliances/$userId/appliances");
-
-    final response = await http.get(url, headers: <String, String>{
-      'Content-Type': 'application/json; charset=UTF-8',
-    });
-
-    if (response.statusCode == 200) {
-      setState(() {
-        appliances = List<Map<String, dynamic>>.from(jsonDecode(response.body));
-        isLoading = false;
-      });
-    } else {
-      print('Failed to fetch appliances: ${response.statusCode}');
-      setState(() {
-        isLoading = false;
-      });
-    }
-  }
-
- */
